Move the basketball open gym component
