+++ src/components/Basketball/displayGyms.js
@@ -1,6 +1,6 @@
 import React from "react";
 import gymsArray from "./gymsArray";
-import { OpenGym } from "../../OpenGymInst";
+import { OpenGym } from "./OpenGymInst";
 
 export default class DisplayGyms extends React.Component {
   render() {
